FILE: nests mock-data and creates barrel file dedicated to mock-data at top level
