+++ libs/core/util-selection/src/lib/mock-data.ts
@@ -1,91 +1,1 @@
-import { Ids } from './internal/ids';
-import { States } from './internal/types';
-
-export interface TestItem {
-  id: string;
-  children?: TestItem[];
-  targets?: number;
-}
-
-export const AFRICA: TestItem = {
-  id: 'Africa',
-  children: [
-    {
-      id: 'Southern Africa',
-      children: [
-        { id: 'Swaziland', targets: 28 },
-        { id: 'Namibia', targets: 17 },
-      ],
-    },
-    { id: 'Central Africa', targets: 65 },
-    {
-      id: 'Northern Africa',
-      children: [
-        {
-          id: 'Morocco',
-          children: [
-            { id: 'Marrakesh', targets: 9 },
-            { id: 'Fes', targets: 11 },
-          ],
-        },
-      ],
-    },
-  ],
-};
-
-export const SMALL_AFRICA: TestItem = {
-  id: 'Africa',
-  children: [
-    {
-      id: 'Southern Africa',
-      children: [{ id: 'Swaziland', targets: 28 }],
-    },
-    {
-      id: 'Northern Africa',
-      children: [
-        {
-          id: 'Morocco',
-          children: [{ id: 'Fes', targets: 11 }],
-        },
-      ],
-    },
-  ],
-};
-
-export const SOME_SELECTED_STATES: States = {
-  Africa: 'indeterminate',
-  Morocco: 'indeterminate',
-  'Southern Africa': 'indeterminate',
-  Swaziland: 'checked',
-  'Northern Africa': 'indeterminate',
-  Fes: 'checked',
-};
-
-export const SOME_SELECTED_ARRAY_MODEL = ['Swaziland', 'Fes'];
-export const SOME_SELECTED_SET_MODEL = new Set(SOME_SELECTED_ARRAY_MODEL);
-
-export const ALL_SELECTED_STATES: States = {
-  Africa: 'checked',
-  'Southern Africa': 'checked',
-  Swaziland: 'checked',
-  Namibia: 'checked',
-  'Central Africa': 'checked',
-  'Northern Africa': 'checked',
-  Morocco: 'checked',
-  Marrakesh: 'checked',
-  Fes: 'checked',
-};
-
-export const ALL_SELECTED_ARRAY_MODEL = [
-  'Central Africa',
-  'Swaziland',
-  'Namibia',
-  'Marrakesh',
-  'Fes',
-];
-export const ALL_SELECTED_SET_MODEL = new Set(ALL_SELECTED_ARRAY_MODEL);
-
-export const getId = ({ id }: TestItem) => id;
-export const getChildren = ({ children }: TestItem) => children ?? [];
-export const getCounts = ({ targets }: TestItem) => targets ?? 0;
-export const MOCK_IDS = new Ids(AFRICA, getId, getChildren);
+export * from './internal/mock-data';
